Complete rework of the GEM bindings. Changed all 'int' in system interfaces to the real type 'short'. Prefixed some macros in gem.h due to collisions with other libs (object states -> OS_, object flags -> OF_, object colors -> OC_).
Programmers attention: you must adjust your source code to the new
bindings and macros names (for compatibility the old macro names
are available if __GEMLIB_OLDNAMES is defined before gem.h is
included.

+++ gemlib/a_appl_read.c
@@ -2,17 +2,21 @@
 #include "gem.h"
 
 
-int
-appl_read (int ApId, int Length, void *ApPbuff)
+short
+appl_read (short ApId, short Length, void *ApPbuff)
 {
 	aes_intin[0] = ApId;
 	aes_intin[1] = Length;
+	
 	aes_addrin[0] = (long) ApPbuff;
+	
 	aes_control[0] = 11;
 	aes_control[1] = 2;
 	aes_control[2] = 1;
 	aes_control[3] = 1;
 	aes_control[4] = 0;
+	
 	aes (&aes_params);
+	
 	return aes_intout[0];
 }
